Re-yeet actions.ts & only open workspace from button
